pyproject.toml
[tool.poetry]
name = "data_ingestion"
version = "0.1.0"
description = "Ingest videos and curate dataset of images"
authors = ["Ian Yu, Bowen Zhang, David Zhu, Elaine Laie"]

[tool.poetry.dependencies]
python = "^3.9"
mediapipe = "^0.8.10"
imageio = "^2.19.2"
opencv-python = "^4.5.5"
ffmpeg = "^1.4"
numpy = "^1.22.3"
pandas = "^1.4.2"
scikit-learn = "^1.1.0"
python-dotenv = "^0.20.0"
pathlib = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
